requirements.txt
fastapi
iso8601
uvicorn
pydantic
psycopg2-binary
sqlalchemy
python-dotenv

passlib[bcrypt]
python-jose[cryptography]
python-multipart

ulid-py
pytest
httpx